+++ adaptor/typed/typedmarshaler/marshaler.go
@@ -6,3 +6,8 @@ type Marshaler[T any] interface {
 	Marshal(T) ([]byte, error)
 	Unmarshal([]byte) (T, error)
 }
+
+// Zero returns the zero-value of type T.
+func Zero[T any]() (_ T) {
+	return
+}